Change GameObject.sector to byte
because it's being sent as a byte anyway, and a map will only ever
contain about 10 to 20 sectors

+++ Client/GameObject.cs
@@ -31,7 +31,7 @@ namespace ShiftDrive {
 
         public Vector2 position;
         public float facing;
-        public int sector;
+        public byte sector;
 
         public string spritename;
         public Texture2D sprite;
